moving the views.hpp
